requirements.txt
paramiko
cryptography==36.0.2
scp
passlib[bcrypt]
requests[security]
uvicorn
eventlet
websockets
jinja2
aiofiles
openpyxl
fastapi
fastapi-mail==1.0.8
redis
pymongo
ldap3==2.5.0
zipfile36
apscheduler
Markdown
markdownify
more-itertools
coverage
python-gitlab
python-dotenv
python-multipart
python-jose[cryptography]
PyYAML
pytesseract
opencv-python-headless
chinesecalendar

